refactor: unify isAdmin/isModerator imports to single canonical module (#593)
Move isModerator from mod.command.ts into admin-auth.utils.ts alongside isAdmin,
update all call sites to import from the canonical location, and remove the
isAdmin re-export that was forwarding through admin.command.ts.
