+++ src/commands/admin/admin-auth.utils.ts
@@ -1,5 +1,10 @@
-import { PermissionsBitField } from "discord.js";
-import { ACCESS_DENIED_ADMIN, AnyRepliable, safeReply } from "../../functions/InteractionUtils.js";
+import { MessageFlags, PermissionsBitField } from "discord.js";
+import {
+  ACCESS_DENIED_ADMIN,
+  ACCESS_DENIED_MOD,
+  AnyRepliable,
+  safeReply,
+} from "../../functions/InteractionUtils.js";
 import { buildTextReply } from "../../functions/ComponentsV2Utils.js";
 
 export async function isAdmin(interaction: AnyRepliable): Promise<boolean> {
@@ -19,4 +24,31 @@ export async function isAdmin(interaction: AnyRepliable): Promise<boolean> {
   }
 
   return isAdmin;
+}
+
+export async function isModerator(interaction: AnyRepliable): Promise<boolean> {
+  const member: any = (interaction as any).member;
+  const canCheck =
+    member && typeof member.permissionsIn === "function" && interaction.channel;
+  let isMod = canCheck
+    ? member.permissionsIn(interaction.channel).has(PermissionsBitField.Flags.ManageMessages)
+    : false;
+
+  if (!isMod) {
+    const isAdminCheck = canCheck
+      ? member.permissionsIn(interaction.channel).has(PermissionsBitField.Flags.Administrator)
+      : false;
+
+    if (!isAdminCheck) {
+      const denial = {
+        content: ACCESS_DENIED_MOD,
+        flags: MessageFlags.Ephemeral,
+      };
+      await safeReply(interaction, denial as any);
+    } else {
+      isMod = true;
+    }
+  }
+
+  return isMod;
 }
