+++ src/commands/gamedb-admin.command.ts
@@ -53,7 +53,7 @@ import {
   type AutoAcceptResult,
   type AllAcceptStats,
 } from "../services/GamedbAuditService.js";
-import { isAdmin } from "./admin.command.js";
+import { isAdmin } from "./admin/admin-auth.utils.js";
 import Game, { IGame } from "../classes/Game.js";
 import GameSearchSynonym from "../classes/GameSearchSynonym.js";
 import GameSearchSynonymDraft, {
